% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformingData.R
\name{rescalePosition}
\alias{rescalePosition}
\title{Position Normalization}
\usage{
rescalePosition(Df)
}
\arguments{
\item{Df}{The dataframe which has the positions as Position.X and Position.Y}
}
\value{
Returns the dataframe with the normalized positions as Centered_Position.X and Centered_Position.Y as well as Distance_From_Edge and Distance_From_Center
}
\description{
Creates a new column in a data frame which goes from 0 to 1000 in x and y directions to -500 to 500.
Also adds new columns of useful information such as distance from nearest edge and distance from the center.
}
